achieved end goal in react, the list appears and dissappears as supposed to
